modules/nlp/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
spacy
python-dotenv
loguru
